Refactor history component
